requirements_docker.txt
# bcrypt==4.0.1
# blinker==1.7.0
# cachetools==5.3.2
# certifi==2023.7.22
# cffi==1.16.0
# charset-normalizer==3.3.2
# click==8.1.7
# cryptography==42.0.7
# debtcollector==2.5.0
# decorator==5.1.1
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
flask-marshmallow==0.15.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
future==0.18.3
gmi_vault==0.15.3
# greenlet==3.0.1
gssapi==1.8.3
gunicorn==22.0.0
healthcheck==1.3.3
# hvac==1.2.1
# idna==3.4
infoblox-client==0.6.0
# iso8601==2.1.0
# itsdangerous==2.1.2
Jinja2==3.1.2
# junos-eznc==2.6.8
# krb5==0.5.1
ldap3==2.9.1
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mod-wsgi==4.9.4
# msgpack==1.0.7
# napalm==4.1.0
# ncclient==0.6.13
# netaddr==0.9.0
# netifaces==0.11.0
# netmiko==4.2.0
# netutils==1.6.0
# ntc_templates==4.0.1
numpy==1.26.1
openpyxl==3.1.2
#oslo.config==9.2.0
#oslo.context==5.2.0
#oslo.i18n==6.1.0
#oslo.log==5.3.0
#oslo.serialization==5.2.0
#oslo.utils==6.3.0
#packaging==23.2
pandas==2.1.2
redis
# paramiko==3.3.1
# pbr==6.0.0
# pyasn1==0.5.0
# pyasn1-modules==0.3.0
# pycparser==2.21
# pycryptodome==3.19.0
# pyeapi==1.0.2
# pyhcl==0.4.5
# pyinotify==0.9.6
PyJWT==2.8.0
PyMySQL==1.1.1
#PyNaCl==1.5.0
#pyOpenSSL==24.1.0
#pyparsing==3.1.1
#pyserial==3.5
#pyspnego==0.10.2
python-dateutil==2.8.2
python-ldap==3.4.3
python3-ldap==0.9.8.4
#pytz==2023.3.post1
#PyYAML==6.0.1
requests==2.31.0
#requests-kerberos==0.14.0
#rfc3986==2.0.0
scp==0.14.5
six==1.16.0
SQLAlchemy==2.0.31
#stevedore==5.1.0
#textfsm==1.1.3
#transitions==0.9.0
#ttp==0.9.5
#ttp-templates==0.3.5
#typing_extensions==4.8.0
#tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
#wrapt==1.16.0
#yamlordereddictloader==0.4.2